move css file into style folder, clean homepage
