+++ src/Components/FoodContainer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'; 
-import "../style.css"; 
+import "../style/style.css"; 
 import {ReviewRating} from "./ReviewRating"; 
 
 function FoodContainer({foodImage, foodName, rating, totalReviews})
